rename: Home 컴포넌트에서 쓰이는 아토믹한 컴포넌트 파일 위치 변경 - `DrawerRanking`, `RankDrawerCard` 컴포넌트를 `/src/home/components`로 위치 변경 - `Home` 컴포넌트를 `/src/home/pages`로 위치 변경
