+++ src/router.tsx
@@ -4,7 +4,7 @@ import { Layout as DrawerLayout } from './drawer/components/Layout/Layout';
 import MyDrawer from './drawer/pages/MyDrawer';
 import Ranking from './drawer/pages/Ranking';
 import Register from './drawer/pages/Register';
-import { Home } from './home/pages/Home/Home';
+import { Home } from './home/pages/Home';
 import { Search } from './search/Search';
 
 export const Router = () => {
